Correct merge price_map.py + add price_calendar.py
